+++ examples/electron/page.js
@@ -1,4 +1,6 @@
 const electron = require('electron');
+const renderTest = require('./test/render.test.js');
+
 const RxDB = require('../../');
 require('babel-polyfill');
 RxDB.plugin(require('pouchdb-adapter-websql'));
@@ -26,55 +28,70 @@ const heroSchema = {
 console.log('hostname: ' + window.location.hostname);
 const syncURL = 'http://' + window.location.hostname + ':10102/';
 
-let database;
 
 const currentWindow = electron.remote.getCurrentWindow();
 
-RxDB
-    .create({
+let _getDatabase; // cached
+function getDatabase() {
+    if (!_getDatabase) _getDatabase = createDatabase();
+    return _getDatabase;
+}
+
+async function createDatabase() {
+    /**
+     * to check if rxdb works correctly, we run some integration-tests here
+     * if you want to use this electron-example as boilerplate, remove this line
+     */
+    await renderTest();
+
+    const db = await RxDB.create({
         name: 'heroesdb' + currentWindow.custom.dbSuffix, // we add a random timestamp in dev-mode to reset the database on each start
         adapter: 'websql',
         password: 'myLongAndStupidPassword'
-    })
-    .then(db => {
-        console.log('creating hero-collection..');
-        database = db;
-        return db.collection({
-            name: 'heroes',
-            schema: heroSchema
-        });
-    })
-    .then(col => {
-        // sync
-        console.log('starting sync');
-        database.heroes.sync({
-            remote: syncURL + 'hero/'
-        });
-
-        col.find()
-            .sort({
-                name: 1
-            })
-            .$.subscribe(function(heroes) {
-                if (!heroes) {
-                    heroesList.innerHTML = 'Loading..';
-                    return;
-                }
-                console.log('observable fired');
-                console.dir(heroes);
-
-                heroesList.innerHTML = heroes
-                    .map(hero => {
-                        return '<li>' +
-                            '<div class="color-box" style="background:' + hero.color + '"></div>' +
-                            '<div class="name" name="' + hero.name + '">' + hero.name + '</div>' +
-                            '</li>';
-                    })
-                    .reduce((pre, cur) => pre += cur, '');
-            });
     });
 
-addHero = function() {
+    console.log('creating hero-collection..');
+    await db.collection({
+        name: 'heroes',
+        schema: heroSchema
+    });
+
+    console.log('starting sync');
+    db.heroes.sync({
+        remote: syncURL + 'hero/'
+    });
+
+    /**
+     * map the result of the find-query to the heroes-list in the dom
+     */
+    db.heroes.find()
+        .sort({
+            name: 1
+        })
+        .$.subscribe(function(heroes) {
+            if (!heroes) {
+                heroesList.innerHTML = 'Loading..';
+                return;
+            }
+            console.log('observable fired');
+            console.dir(heroes);
+
+            heroesList.innerHTML = heroes
+                .map(hero => {
+                    return '<li>' +
+                        '<div class="color-box" style="background:' + hero.color + '"></div>' +
+                        '<div class="name" name="' + hero.name + '">' + hero.name + '</div>' +
+                        '</li>';
+                })
+                .reduce((pre, cur) => pre += cur, '');
+        });
+
+    return db;
+};
+getDatabase();
+
+window.addHero = async function() {
+    const db = await getDatabase();
     const name = document.querySelector('input[name="name"]').value;
     const color = document.querySelector('input[name="color"]').value;
     const obj = {
@@ -83,5 +100,5 @@ addHero = function() {
     };
     console.log('inserting hero:');
     console.dir(obj);
-    database.heroes.insert(obj);
+    db.heroes.insert(obj);
 };